requirements.txt
--index-url https://download.pytorch.org/whl/cu124
--extra-index-url https://pypi.org/simple

torch==2.5.1
numpy==2.1.3
matplotlib==3.9.2
PyYAML==6.0.2
tqdm==4.67.1
wandb==0.19.11
Pillow==11.1.0
h5py==3.12.1
